ROCm:  MDRange unit tests working

+++ core/src/ROCm/Kokkos_ROCm_Exec.hpp
@@ -205,9 +205,10 @@ struct ROCmParallelLaunch< DriverType
 
       // Invoke the driver function on the device
       auto ext = hc::extent<3>(grid.z,grid.y,grid.x);
-      int bx = (grid.x > block.x)? block.x : grid.x;
-      int by = (grid.y > block.y)? block.y : grid.y;
-      int bz = (grid.z > block.z)? block.z : grid.z;
+      size_t bx = (grid.x > block.x)? block.x : grid.x;
+      size_t by = (grid.y > block.y)? block.y : grid.y;
+      size_t bz = (grid.z > block.z)? block.z : grid.z;
+
       hc::parallel_for_each(ext.tile_with_dynamic(bz,by,bx,shmem), [=](const hc::index<3> & idx) [[hc]]
  
       { rocm_memory_buffer->operator()();
@@ -245,9 +246,9 @@ struct ROCmParallelLaunch< DriverType
 //      KOKKOS_ENSURE_ROCM_LOCK_ARRAYS_ON_DEVICE();
       // Invoke the driver function on the device
       auto ext = hc::extent<3>(grid.z,grid.y,grid.x);
-      int bx = (grid.x > block.x)? block.x : grid.x;
-      int by = (grid.y > block.y)? block.y : grid.y;
-      int bz = (grid.z > block.z)? block.z : grid.z;
+      size_t bx = (grid.x > block.x)? block.x : grid.x;
+      size_t by = (grid.y > block.y)? block.y : grid.y;
+      size_t bz = (grid.z > block.z)? block.z : grid.z;
 
       hc::parallel_for_each(ext.tile_with_dynamic(bz,by,bx,shmem), [=](const hc::index<3> & idx) [[hc]]
